debrisPredict/requirements.txt
flask==2.3.3
gunicorn==21.2.0
numpy>=1.21.0
Pillow>=8.3.1
requests>=2.26.0
python-dotenv>=0.19.0
